# book-network/src/main/java/com/sammy/book_network/book/BorrowedBookResponse.java
package com.sammy.book_network.book;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {
    // this is a combination of bookHistory and Book entity
    private int id;
    private String title;
    private String authorName;
    private String isbn;
    private double rate;
    private boolean returned;
    private boolean returnApproved;
}